+++ CHANGELOG.md
@@ -45,6 +45,7 @@
 - refactor(core): replace uuid4 with service path as the unique identifier of ubo services
 - refactor(core): introduce `utils.dataclass.default_provider` and use it wherever we had that duplicated code for setting the default value of a dataclass field based on other fields
 - refactor: replace all `ApplicationItem`s with the new `UboApplicationItem`, not storing the mutable `PageWidget`  classes in the store, an step towards implementing #261
+- feat(file-system): add `file-system` service with readonly features like navigation through file-system, checking size/permission/owner/group of files and directories, and reading files - closes #299
 
 ## Version 1.3.0